Revert "Merging combined view"
